fix linker errors related to 'inline' when compiling with clang
There were a few function prototypes that missed the 'inline' specifier
that is specified in the actual function definition. Clang chose not to
inline some of these functions and to instead rely on an external
definitions (indicated by the non-inline prototypes), resulting in a
linker error.

I've fixed the issue by adding inline to the prototypes as well. It
might be better to simply not use inline, though.